+++ src/i18n.lib.js
@@ -8,4 +8,4 @@ import { TT, loadI18NResource } from '@nuogz/i18n';
 loadI18NResource('@nuogz/aegis', resolve(dirname(fileURLToPath(import.meta.url)), '..', 'locale'));
 
 
-export const T = TT('@nuogz/aegis');
+export const { T, TS } = TT('@nuogz/aegis');